backend/tempo_api.py
"""
NASA TEMPO satellite data integration
"""
from fastapi import APIRouter, HTTPException, Query
import httpx
import os
from typing import List, Optional
from pydantic import BaseModel
from datetime import datetime
import asyncio

router = APIRouter()

class TempoDataPoint(BaseModel):
    latitude: float
    longitude: float
    value: float
    parameter: str
    timestamp: str

class TempoGridResponse(BaseModel):
    parameter: str
    data: List[TempoDataPoint]
    timestamp: str
    bounds: dict

@router.get("/api/tempo/grid", response_model=TempoGridResponse)
async def get_tempo_grid_data(
    parameter: str = Query("no2", description="Parameter to fetch (no2, o3)"),
    lat_min: float = Query(-90, description="Minimum latitude"),
    lat_max: float = Query(90, description="Maximum latitude"),
    lon_min: float = Query(-180, description="Minimum longitude"),
    lon_max: float = Query(180, description="Maximum longitude"),
):
    """
    Get TEMPO satellite data as a grid for heatmap visualization
    
    Note: This is a simplified implementation. Real TEMPO data requires
    more complex handling of NetCDF/HDF5 files from NASA's data portals.
    
    For production, you would:
    1. Access NASA Earthdata: https://earthdata.nasa.gov/
    2. Use TEMPO L2 products from: https://tempo.si.edu/
    3. Process NetCDF files with xarray/netCDF4
    """
    
    nasa_api_key = os.getenv("NASA_API_KEY", "DEMO_KEY")
    
    try:
        # For demonstration, generate synthetic TEMPO-like data
        # In production, fetch real data from NASA APIs
        grid_data = generate_synthetic_tempo_grid(
            parameter, lat_min, lat_max, lon_min, lon_max
        )
        
        return TempoGridResponse(
            parameter=parameter,
            data=grid_data,
            timestamp=datetime.utcnow().isoformat(),
            bounds={
                "lat_min": lat_min,
                "lat_max": lat_max,
                "lon_min": lon_min,
                "lon_max": lon_max
            }
        )
        
    except Exception as e:
        raise HTTPException(
            status_code=500, 
            detail=f"Failed to fetch TEMPO data: {str(e)}"
        )

def generate_synthetic_tempo_grid(
    parameter: str,
    lat_min: float,
    lat_max: float,
    lon_min: float,
    lon_max: float,
    resolution: float = 0.5
) -> List[TempoDataPoint]:
    """
    Generate synthetic TEMPO-like satellite data for visualization
    
    Real implementation would:
    - Download TEMPO L2 data products
    - Parse NetCDF/HDF5 files
    - Regrid to desired resolution
    - Apply quality filters
    """
    import numpy as np
    
    # Create grid
    lats = np.arange(lat_min, lat_max, resolution)
    lons = np.arange(lon_min, lon_max, resolution)
    
    data_points = []
    
    # Generate synthetic pollution patterns
    for lat in lats:
        for lon in lons:
            # Create realistic patterns with hotspots near urban areas
            # Higher values near equator and major cities
            base_value = 15.0
            
            # Add some variation based on location
            lat_factor = 1.0 + 0.3 * np.sin(np.radians(lat * 4))
            lon_factor = 1.0 + 0.2 * np.cos(np.radians(lon * 3))
            
            # Add random noise
            noise = np.random.normal(0, 3)
            
            if parameter == "no2":
                # NO2 typically ranges 0-100 (×10^15 molecules/cm²)
                value = max(0, base_value * lat_factor * lon_factor + noise)
            elif parameter == "o3":
                # O3 typically ranges 20-80 DU (Dobson Units)
                value = max(20, 45 + 10 * lat_factor * lon_factor + noise)
            else:
                value = 0
            
            # Only include points with significant values
            if value > 5:
                data_points.append(TempoDataPoint(
                    latitude=float(lat),
                    longitude=float(lon),
                    value=float(value),
                    parameter=parameter,
                    timestamp=datetime.utcnow().isoformat()
                ))
    
    return data_points

@router.get("/api/tempo/parameters")
async def get_available_parameters():
    """Get list of available TEMPO parameters"""
    return {
        "parameters": [
            {
                "id": "no2",
                "name": "Nitrogen Dioxide",
                "unit": "×10^15 molecules/cm²",
                "description": "Tropospheric NO₂ column density"
            },
            {
                "id": "o3",
                "name": "Ozone",
                "unit": "DU (Dobson Units)",
                "description": "Total column ozone"
            }
        ]
    }

# Instructions for real TEMPO data integration:
"""
REAL TEMPO DATA INTEGRATION STEPS:

1. Register for NASA Earthdata account:
   https://urs.earthdata.nasa.gov/users/new

2. Access TEMPO data portal:
   https://asdc.larc.nasa.gov/project/TEMPO

3. Install required libraries:
   pip install netCDF4 xarray h5py

4. Example code for real data:

import xarray as xr
import numpy as np

async def fetch_real_tempo_data(date, parameter):
    # Download TEMPO L2 file
    file_path = await download_tempo_l2_file(date, parameter)
    
    # Open NetCDF file
    ds = xr.open_dataset(file_path)
    
    # Extract data
    if parameter == "no2":
        data = ds['vertical_column_troposphere']
        lat = ds['latitude']
        lon = ds['longitude']
    
    # Filter quality
    quality_flag = ds['quality_flag']
    data = data.where(quality_flag > 0.5)
    
    # Convert to grid
    grid_data = []
    for i in range(len(lat)):
        for j in range(len(lon)):
            if not np.isnan(data[i, j]):
                grid_data.append({
                    'latitude': float(lat[i, j]),
                    'longitude': float(lon[i, j]),
                    'value': float(data[i, j])
                })
    
    return grid_data

5. Update the endpoint to use real data instead of synthetic
"""

